move useInnerText hook
